# front/openapi.yaml
openapi: 3.1.0
info:
  title: FerrisKey API
  description: FerrisKey API
  contact:
    name: nathaelb
    email: pro.nathaelbonnal@gmail.com
  license:
    name: ""
  version: 0.1.0
paths:
...
  /realms/{realm_name}/protocol/openid-connect/logout:
    get:
      tags:
        - auth
      summary: OIDC RP-Initiated Logout
      description: Ends the user's OP session. Supports id_token_hint, post_logout_redirect_uri, state, and client_id.
      operationId: logout_get
      parameters:
        - name: realm_name
          in: path
          description: Realm name
          required: true
          schema:
            type: string
        - name: id_token_hint
          in: query
          required: false
          schema:
            type: string
        - name: post_logout_redirect_uri
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: client_id
          in: query
          required: false
          schema:
            type: string
      responses:
        "204":
          description: Session cookies cleared
        "307":
          description: Redirect to post_logout_redirect_uri
        "400":
          description: Invalid or missing logout parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
        - auth
      summary: OIDC RP-Initiated Logout
      description: Ends the user's OP session. Supports id_token_hint, post_logout_redirect_uri, state, and client_id.
      operationId: logout
      parameters:
        - name: realm_name
          in: path
          description: Realm name
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/LogoutRequestValidator"
      responses:
        "204":
          description: Session cookies cleared
        "307":
          description: Redirect to post_logout_redirect_uri
        "400":
          description: Invalid or missing logout parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - code
        - status
        - message
      properties:
        code:
          type: string
        status:
          type: integer
          format: int32
        message:
          type: string
...
    LogoutRequestValidator:
      type: object
      properties:
        client_id:
          type:
            - string
            - "null"
        id_token_hint:
          type:
            - string
            - "null"
        post_logout_redirect_uri:
          type:
            - string
            - "null"
        state:
          type:
            - string
            - "null"
...